[Lutron] Move bridge state change handling to LutronHandler
